backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.26
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.28.1
pyjwt==2.10.1
cryptography==45.0.2